feat: moved more functions to separate node modules
