+++ lib/apply-group-by.js
@@ -1,0 +1,42 @@
+const async = require('async');
+const Q = require('q');
+
+/**
+ * TODO: Group by.
+ *
+ * @param {Array} imagesToGroup - Array of image objects.
+ * @param {object} opts - Options passed to the plugin.
+ * @returns {Promise}
+ */
+function applyGroupBy(imagesToGroup, opts) {
+  return Q.Promise((resolve, reject) => {
+    async.reduce(
+      opts.groupBy,
+      imagesToGroup,
+      (images, group, next) => {
+        async.map(
+          images,
+          (image, done) => {
+            new Q(group(image))
+              .then((groupedImage) => {
+                if (groupedImage) {
+                  image.groups.push(groupedImage);
+                }
+                done(null, image);
+              })
+              .catch(done);
+          },
+          next
+        );
+      },
+      (err, images) => {
+        if (err) {
+          return reject(err);
+        }
+        return resolve([images, opts]);
+      }
+    );
+  });
+}
+
+exports.applyGroupBy = applyGroupBy;
