+++ lib/log.js
@@ -7,7 +7,6 @@ const fancyLog = require('fancy-log');
  *
  */
 function log(...message) {
-  // const data = Array.prototype.slice.call(arguments);
   fancyLog.apply(false, message);
 }
