+++ lib/mask.js
@@ -9,6 +9,7 @@ const { GROUP_DELIMITER, GROUP_MASK } = require('./constants');
 function mask(toggle) {
   const input = new RegExp(`[${toggle ? GROUP_DELIMITER : GROUP_MASK}]`, 'gi');
   const output = toggle ? GROUP_MASK : GROUP_DELIMITER;
+
   return (value) => {
     return value.replace(input, output);
   };
