+++ lib/retina-images.js
@@ -1,5 +1,3 @@
-const lodash = require('lodash');
-
 /**
  * Check whether the image is retina.
  *
@@ -7,7 +5,9 @@ const lodash = require('lodash');
  * @returns {boolean} Whether the image is retina.
  */
 function isRetinaImage(imageUrl) {
-  return /@(\d)x\.[a-z]{3,4}$/gi.test(imageUrl.split('#')[0]);
+  const retinaPattern = /@(\d)x\.[a-z]{3,4}$/gi;
+
+  return retinaPattern.test(imageUrl.split('#')[0]);
 }
 
 /**
@@ -17,8 +17,10 @@ function isRetinaImage(imageUrl) {
  * @returns {number} The retina ratio.
  */
 function getRetinaRatio(imageUrl) {
+  const retinaPattern = /@(\d)x\.[a-z]{3,4}$/gi;
+
   // Find any @{number}x matches in the url string.
-  const matches = /@(\d)x\.[a-z]{3,4}$/gi.exec(imageUrl.split('#')[0]);
+  const matches = retinaPattern.exec(imageUrl.split('#')[0]);
   const ratio = parseInt(matches[1], 10);
 
   return ratio;
@@ -28,10 +30,10 @@ function getRetinaRatio(imageUrl) {
  * Check whether all images are retina.
  *
  * @param {Array} images - The images to check.
- * @returns {boolean} Whether the images are all retina.
+ * @returns {boolean} 'true' if the images are all retina.
  */
 function areAllRetina(images) {
-  return lodash.every(images, (image) => {
+  return images.every((image) => {
     return image.ratio > 1;
   });
 }
