+++ lib/sprites.js
@@ -4,7 +4,6 @@ const path = require('path');
 const fs = require('fs');
 const mkdirp = require('mkdirp');
 const ansi = require('ansi-colors');
-
 const { log } = require('./log');
 
 /**
