+++ lib/tokens.js
@@ -1,6 +1,6 @@
 const Q = require('q');
-const lodash = require('lodash');
 const postcss = require('postcss');
+
 const { getBackgroundColor } = require('./get-background-values');
 const { hasImageInRule, getImageUrl } = require('./image-urls');
 const { BACKGROUND, BACKGROUND_IMAGE } = require('./constants');
@@ -36,7 +36,9 @@ function setTokens(images, opts, css) {
       // in them.
       if (hasImageInRule(rule.toString())) {
         url = getImageUrl(rule.toString());
-        image = lodash.find(images, { url });
+        image = images.find((img) => {
+          return img.url === url;
+        });
 
         if (image) {
           // We remove these declarations since
